chore: Rename MonthCalendar semantics with child components
