+++ packages/components/src/DateTimePickers/DatePicker/CalendarBody/index.js
@@ -1,0 +1,3 @@
+import CalendarBody from './CalendarBody.component';
+
+export default CalendarBody;
