+++ packages/components/src/DateTimePickers/DatePicker/CalendarHeader/CalendarHeader.scss
