+++ packages/components/src/DateTimePickers/DatePicker/CalendarHeader/CalendarHeader.test.js
@@ -1,0 +1,7 @@
+import React from 'react';
+import { shallow } from 'enzyme';
+
+import CalendarHeader from './CalendarHeader.component';
+
+describe('CalendarHeader', () => {
+});
