+++ packages/components/src/DateTimePickers/DatePicker/CalendarHeader/index.js
@@ -1,0 +1,3 @@
+import CalendarHeader from './CalendarHeader.component';
+
+export default CalendarHeader;
