+++ packages/components/src/DateTimePickers/DatePicker/DatePicker.scss
