+++ packages/components/src/DateTimePickers/DatePicker/DatePicker.test.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import { shallow } from 'enzyme';
+
+import DatePicker from './DatePicker.component';
+
+describe('DatePicker', () => {
+
+});
