+++ packages/components/src/DateTimePickers/DatePicker/index.js
@@ -1,0 +1,3 @@
+import DatePicker from './DatePicker.component';
+
+export default DatePicker;
