--- packages/components/src/DateTimePickers/MonthCalendar/MonthCalendar.test.js
@@ -1,8 +1,0 @@
-import React from 'react';
-import { shallow } from 'enzyme';
-
-import MonthCalendar from './MonthCalendar.component';
-
-describe('MonthCalendar', () => {
-
-});
