--- packages/components/src/DateTimePickers/MonthCalendar/MonthCalendarDays/index.js
@@ -1,3 +1,0 @@
-import MonthCalendarDays from './MonthCalendarDays.component';
-
-export default MonthCalendarDays;
