--- packages/components/src/DateTimePickers/MonthCalendar/MonthCalendarHeader/MonthCalendarHeader.test.js
@@ -1,7 +1,0 @@
-import React from 'react';
-import { shallow } from 'enzyme';
-
-import MonthCalendarHeader from './MonthCalendarHeader.component';
-
-describe('MonthCalendarHeader', () => {
-});
