--- packages/components/src/DateTimePickers/MonthCalendar/MonthCalendarHeader/index.js
@@ -1,3 +1,0 @@
-import MonthCalendarHeader from './MonthCalendarHeader.component';
-
-export default MonthCalendarHeader;
