--- packages/components/src/DateTimePickers/MonthCalendar/index.js
@@ -1,3 +1,0 @@
-import MonthCalendar from './MonthCalendar.component';
-
-export default MonthCalendar;
